move the shortuuid code to shortuuid.py
