+++ spylib/utils/__init__.py
@@ -2,9 +2,10 @@ from .domain import domain_to_storename, store_domain
 from .hmac import validate
 from .httpclient import HTTPClient
 from .jwtoken import JWTBaseModel
-from .misc import get_unique_id, now_epoch
+from .misc import now_epoch
 from .rest import DELETE, GET, POST, PUT, Method
 from .session_token import SessionToken
+from .shortuuid import get_unique_id
 
 __all__ = [
     'now_epoch',
