--- spylib/utils/constants.py
@@ -1,1 +1,0 @@
-ALPHABET = '23456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz'
